feat: implement thin httpx SPARQLWrapper
All SPARQL query execution in RDFProxy can run in a thin SPARQLWrapper
class based on httpx. This allows to discontinue the use of the
SPARQLWrapper library.

Closes #208.
