+++ rdfproxy/__init__.py
@@ -1,8 +1,4 @@
 from rdfproxy.adapter import SPARQLModelAdapter  # noqa: F401
-from rdfproxy.sparql_strategies import (
-    SPARQLStrategy,  # noqa: F401
-    SPARQLWrapperStrategy,  # noqa: F401
-    HttpxStrategy,  # noqa: F401
-)
+from rdfproxy.sparqlwrapper import SPARQLWrapper  # noqa: F401
 from rdfproxy.utils._types import ConfigDict, SPARQLBinding  # noqa: F401
 from rdfproxy.utils.models import Page, QueryParameters  # noqa: F401
